Begin rewriting gspreadtest, refining main.py
This update will start a major rewrite of the core of gspreadtest, the main
building-block of the spreadsheet operations. It will also start refining
the code for the databot itself, which depends on gspreadtest.

The new variant of "gspreadtest" is designed to follow the original
specification as closely as possible (using the same method names whenever
possible), but it will instead use ODS files and Nextcloud as opposed to
Google Drive documents.

Obviously, more work is needed, but the reader will notice that the basic test
depicted in the new gspreadtest file should be functional.

Let us be free and control our own data!