Use and check for proper .ods filename in get_book
This update will serve as a proper "glue" between the spreadsh_mgr library and
the main bot routine... and it should now enable the `;displayHeaders` command
to work (at least, the first time it is run). More debugging is pending.

The update will also add other stability improvements, like properly setting the
default worksheet name for new books and invoking the `0` index to get the first
row instead of using the `1` index.