Start to generalize methods for many guilds/sheets
This update will expand the write_cell, get_row, and other methods in order to
specify the file name and the sheet name. This should help avoid aforementioned
race conditions. It will also solve a logical error in displayHeaders, so that
method and the newHeader method should now work successfully.